Extract some stuff into `LoadingEndMgr`
